(Tech) refactor(accessibilityAndTestId): move outside test libs and update imports+snapshots
