+++ src/features/auth/signup/PhoneValidation/SetPhoneNumber.tsx
@@ -10,6 +10,7 @@ import { QuitSignupModal } from 'features/auth/components/QuitSignupModal'
 import { useAppSettings } from 'features/auth/settings'
 import { SignupStep } from 'features/auth/signup/enums'
 import { UseNavigationType } from 'features/navigation/RootNavigator'
+import { accessibilityAndTestId } from 'libs/accessibilityAndTestId'
 import { amplitude } from 'libs/amplitude'
 import { CountryPicker, METROPOLITAN_FRANCE } from 'libs/country-picker'
 import { currentTimestamp } from 'libs/dates'
@@ -17,7 +18,6 @@ import { useSafeState } from 'libs/hooks'
 import { captureMonitoringError } from 'libs/monitoring'
 import { storage } from 'libs/storage'
 import { TIMER_NOT_INITIALIZED, useTimer } from 'libs/timer'
-import { accessibilityAndTestId } from 'tests/utils'
 import { BottomContentPage } from 'ui/components/BottomContentPage'
 import { ButtonPrimary } from 'ui/components/buttons/ButtonPrimary'
 import { InputError } from 'ui/components/inputs/InputError'
