+++ src/features/identityCheck/pages/profile/SetAddress.tsx
@@ -11,9 +11,9 @@ import { PageWithHeader } from 'features/identityCheck/components/layout/PageWit
 import { useIdentityCheckContext } from 'features/identityCheck/context/IdentityCheckContextProvider'
 import { IdentityCheckError } from 'features/identityCheck/errors'
 import { useIdentityCheckNavigation } from 'features/identityCheck/useIdentityCheckNavigation'
+import { accessibilityAndTestId } from 'libs/accessibilityAndTestId'
 import { eventMonitoring } from 'libs/monitoring'
 import { useAddresses } from 'libs/place'
-import { accessibilityAndTestId } from 'tests/utils'
 import { ButtonPrimary } from 'ui/components/buttons/ButtonPrimary'
 import { isAddressValid } from 'ui/components/inputs/addressCheck'
 import { InputError } from 'ui/components/inputs/InputError'
