+++ src/features/profile/pages/Profile.tsx
@@ -14,11 +14,11 @@ import { ProfileHeader } from 'features/profile/components/ProfileHeader'
 import { ProfileContainer } from 'features/profile/components/reusables'
 import { SectionWithSwitch } from 'features/profile/components/SectionWithSwitch'
 import { SmartBanner } from 'features/smartBanner/SmartBanner'
+import { accessibilityAndTestId } from 'libs/accessibilityAndTestId'
 import { analytics, isCloseToBottom } from 'libs/analytics'
 import { env } from 'libs/environment'
 import { GeolocPermissionState, useGeolocation } from 'libs/geolocation'
 import useFunctionOnce from 'libs/hooks/useFunctionOnce'
-import { accessibilityAndTestId } from 'tests/utils'
 import { InputError } from 'ui/components/inputs/InputError'
 import { Section } from 'ui/components/Section'
 import { SectionRow } from 'ui/components/SectionRow'
