+++ src/libs/accessibilityAndTestId.ts
@@ -1,0 +1,7 @@
+export function accessibilityAndTestId(accessibilityLabel?: string, testID?: string) {
+  return {
+    accessible: !!accessibilityLabel,
+    accessibilityLabel,
+    testID: testID || accessibilityLabel,
+  }
+}
