+++ src/libs/accessibilityAndTestId.web.ts
@@ -1,0 +1,3 @@
+export function accessibilityAndTestId(id: string) {
+  return { accessible: true, accessibilityLabel: id, testID: id }
+}
